One tag create and remove tests
